Re-added functions and outputs files kept some methods in backend
